pyproject.toml
[tool.poetry]
name = "conf-ark"
version = "0.1.0"
description = "A tool to backup and restore config files from user directory"
authors = ["Slava Yozhkin <yozhkin@gmail.com>"]
readme = "README.md"
packages = [{include = "conf_ark"}]

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.1.7"
pathlib = "^1.0.1"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"

[tool.poetry.scripts]
conf-ark = "conf_ark.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
